修改

+++ OperationalApplicationSystem/src/main/webapp/UI/sign/stamp/stampform_all_usedFile.jsp
@@ -48,9 +48,12 @@ function deleteFile(path,o){
 	
 }		
 
-function fileToTr(name,path,b){
-	return "<tr><td align='center'><a onclick=\"getFile('"+path.replace('\\','\\\\')+"')\" url='"+path.replace('\\','\\\\')+ "' >"+name+"</></td><td align='center'><a onclick=\"deleteFile('"+path.replace('\\','\\\\')+"',this)\" >Delete</a></td></tr>"
-
+function fileToTr(name,path){
+	if('${user.name}'=='admin'){
+		return "<tr><td align='center'><a onclick=\"getFile('"+path.replace('\\','\\\\')+"')\" url='"+path.replace('\\','\\\\')+ "' >"+name+"</></td><td align='center'><a onclick=\"deleteFile('"+path.replace('\\','\\\\')+"',this)\" >Delete</a></td></tr>"
+	}else{
+		return "<tr><td align='center'><a onclick=\"getFile('"+path.replace('\\','\\\\')+"')\" url='"+path.replace('\\','\\\\')+ "' >"+name+"</></td><td align='center'></td></tr>"
+	}
 }
 
 function usedFileUpload(id,tid) {
